+++ brayns/common/Statistics.h
@@ -28,13 +28,13 @@ namespace brayns
 class Statistics : public BaseObject
 {
 public:
-    double getFPS() const;
-    void setFPS(const double fps);
+    uint32_t getFPS() const;
+    void setFPS(const uint32_t fps);
     size_t getSceneSizeInBytes() const;
     void setSceneSizeInBytes(const size_t sceneSizeInBytes);
 
 private:
-    double _fps{0.0};
+    uint32_t _fps{0};
     size_t _sceneSizeInBytes{0};
 };
 } // namespace brayns
